[Refact] 설문 도메인 관리자/일반 분리 and DTO 필드명 수정 (#196)
* [refact] #189 dto 필드명 공통화

* [refact] #189 dto 생성 로직을 service에 퍼져있는것을 dto 클래스로 옮김
- 필드명을 고칠 때, service 로직까지 고쳐줘야 하는게 문제가 있다고 판단함

* [refact] #189 survey 관련 기능 일반과 관리자 기능으로 controller 분리
