+++ src/main/java/com/abcdedu_backend/survey/dto/response/SurveyListResponse.java
@@ -8,7 +8,16 @@ import java.time.LocalDateTime;
 public record SurveyListResponse (
     Long id,
     String title,
-    LocalDateTime createAt,
+    LocalDateTime createdAt,
     String writerName
 ){
+    public static SurveyListResponse of(Long id, String title, String writerName, LocalDateTime createdAt) {
+        return SurveyListResponse.builder()
+                .id(id)
+                .title(title)
+                .writerName(writerName)
+                .createdAt(createdAt)
+                .build();
+    }
+
 }
